pyproject.toml
[tool.poetry]
name = "cs2-battle-bot"
version = "0.2.0"
description = ""
authors = ["Adrian Ciolek <ciolek.adrian@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
django-prefix-id = "^1.0.0"
steam = "^1.4.4"
django-redis = "^5.4.0"
djangorestframework = "^3.15.2"
djangorestframework-api-key = "^3.0.0"
rcon = "^2.4.6"
gunicorn = "^22.0.0"
drf-spectacular = "^0.27.2"
drf-standardized-errors = "^0.14.0"
django-group-model = "^1.0.1"
py-cord = "^2.5.0"
django-allauth = {extras = ["socialaccount", "steam"], version = "^0.63.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
